Refactored code structure to more go-like structure
- Split state.go into store.go and cache.go
- Altered interfaces to just be redis client specific

Resolves FAPI-2242
